requirements.txt
openai>=1.12.0
anthropic>=0.18.0
transformers>=4.36.0
torch>=2.1.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
gTTS>=2.5.0
pyaudio>=0.2.14
opencv-python>=4.9.0
mediapipe>=0.10.9
deepface>=0.0.79
nltk>=3.8.1
textblob>=0.17.1
vaderSentiment>=3.3.2
PyPDF2>=3.0.0
python-docx>=1.1.0
numpy>=1.24.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
reportlab>=4.0.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
requests>=2.31.0
Werkzeug>=2.3.0
Flask>=2.3.0
Flask-CORS>=4.0.0
gunicorn>=21.2.0